Made putnbr more readable

+++ libftprintf.h
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_printf.h                                        :+:      :+:    :+:   */
+/*   libftprintf.h                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: msimoes <msimoes@student.42lisboa.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/28 10:30:07 by msimoes           #+#    #+#             */
-/*   Updated: 2025/04/29 13:03:46 by msimoes          ###   ########.fr       */
+/*   Updated: 2025/04/29 15:36:08 by msimoes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 
 int	ft_putchar(int c);
 int	ft_putstr(char *str);
-int	ft_putnbr(int n);
+int	ft_putnbr(long n);
 int	ft_unsputnbr(unsigned int n);
 
 #endif
